trying some ideas for checking moves

+++ chess/src/main.c
@@ -195,7 +195,7 @@ bool cursor_action( CURSOR* cursor, CHESS_PIECE brd[8][8], u8 player ) {
         // A piece is currently selected, check if cursor position is valid )
         //TODO -- (add valid move check, for now just look for empty squares)
         // if( brd[(u8)cursor->col][(u8)cursor->row].type == EMPTY ) { 
-        if( is_valid_move( (u8) cursor->sel_col, (u8) cursor->sel_row,(u8)  cursor->col,(u8)  cursor->row ) ){
+        if( is_valid_move( cursor->sel_col, cursor->sel_row, cursor->col, cursor->row ) ){
             move_piece( cursor->sel_col, cursor->sel_row, cursor->col, cursor->row );
             return true;
         } else {
